+++ azext_iot/central/providers/device_provider.py
@@ -121,6 +121,44 @@ class CentralDeviceProvider:
 
         return self._devices[device.id]
 
+    def update_device(
+        self,
+        device_id,
+        device_name=None,
+        template=None,
+        simulated=None,
+        enabled=None,
+        organizations=None,
+        central_dns_suffix=CENTRAL_ENDPOINT,
+    ) -> Union[DeviceV1, DeviceV1_1_preview, DevicePreview]:
+        if not device_id:
+            raise CLIError("Device id must be specified.")
+
+        if device_id in self._devices:
+            raise CLIError("Device already exists.")
+
+        device = central_services.device.update_device(
+            cmd=self._cmd,
+            app_id=self._app_id,
+            device_id=device_id,
+            device_name=device_name,
+            template=template,
+            simulated=simulated,
+            enabled=enabled,
+            organizations=organizations,
+            token=self._token,
+            central_dns_suffix=central_dns_suffix,
+            api_version=self._api_version,
+        )
+
+        if not device:
+            raise CLIError("No device found with id: '{}'.".format(device_id))
+
+        # add to cache
+        self._devices[device.id] = device
+
+        return self._devices[device.id]
+
     def delete_device(
         self,
         device_id,
@@ -186,7 +224,7 @@ class CentralDeviceProvider:
             return info
 
         device = self.get_device(device_id, central_dns_suffix)
-        if device.device_status == DeviceStatus.provisioned:
+        if device._device_status == DeviceStatus.provisioned:
             credentials = self.get_device_credentials(
                 device_id=device_id,
                 central_dns_suffix=central_dns_suffix,
@@ -196,7 +234,7 @@ class CentralDeviceProvider:
             dps_state = dps_global_service.get_registration_state(
                 id_scope=id_scope, key=key, device_id=device_id
             )
-        dps_state = self._dps_populate_essential_info(dps_state, device.device_status)
+        dps_state = self._dps_populate_essential_info(dps_state, device._device_status)
 
         info = {
             "@device_id": device_id,