+++ azext_iot/tests/central/__init__.py
@@ -345,7 +345,7 @@ class CentralLiveScenarioTest(CaptureOutputLiveScenarioTest):
                 self.check("account", None if account_name is None else account_name),
                 self.check("state", "pending"),
                 self.check(
-                    "sasttl", DEFAULT_FILE_UPLOAD_TTL if sasttl is None else sasttl
+                    "sasTtl", DEFAULT_FILE_UPLOAD_TTL if sasttl is None else sasttl
                 ),
             ],
         ).get_output_in_json()